+++ test/bigtest/tests/units/settings/JopProfileDetails/JobProfileDetails-test.js
@@ -174,9 +174,11 @@ describe('JobProfileDetails', () => {
           expect(deletingConfirmationModal.confirmButton.text).to.equal(translations.delete);
         });
 
-        describe('clicking on cancel button', () => {
+        // TODO: Enable tests after resolving STCOR-445
+        describe.skip('clicking on cancel button', () => {
           beforeEach(async () => {
             await deletingConfirmationModal.cancelButton.click();
+            await wait(2000);
           });
 
           it('should hide delete confirmation modal', () => {