+++ test/bigtest/tests/units/settings/MappingProfilesForm/interactors/MappingProfilesFormInteractor.js
@@ -4,9 +4,9 @@ import ExpandAllButtonInteractor from '@folio/stripes-components/lib/Accordion/t
 import ButtonInteractor from '@folio/stripes-components/lib/Button/tests/interactor';
 import { AccordionSetInteractor } from '@folio/stripes-components/lib/Accordion/tests/interactor';
 import { FullScreenFormInteractor } from '@folio/stripes-data-transfer-components/interactors';
+import MultiColumnListInteractor from '@folio/stripes-components/lib/MultiColumnList/tests/interactor';
 
 import { SummaryContentInteractor } from './SummaryContentInteractor';
-import { TransformationsInteractor } from './TransformationsInteractor';
 import { TransformationsModalInteractor } from './TransformationsModalInteractor';
 
 @interactor
@@ -18,6 +18,6 @@ export class MappingProfilesFormInteractor {
   addTransformationButton = new ButtonInteractor('[data-test-add-transformation]');
   accordions = new AccordionSetInteractor('#mapping-profiles-form-accordions');
   summary = new SummaryContentInteractor();
-  transformations = new TransformationsInteractor();
+  transformations = new MultiColumnListInteractor('#mapping-profile-transformations-list');
   transformationsModal = new TransformationsModalInteractor();
 }